attributes and migrations being changed

+++ database/migrations/2024_11_25_164650_create_schedules_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('schedules', function (Blueprint $table) {
             $table->id();
-            $table->integer('day_id');
+            $table->unsignedBigInteger('day_id');
             $table->string('title');
             $table->text('description');
             $table->string('color')->nullable();
